% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BibEntryExtractOp.R
\name{FindBibEntry}
\alias{FindBibEntry}
\title{Find a search term in the specified field of a BibEntry object}
\usage{
FindBibEntry(bib, term, field)
}
\description{
Workhorse function for SearchBib
}
\keyword{internal}

