% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MatchName}
\alias{MatchName}
\title{Search for a name in a BibEntry object}
\usage{
MatchName(nom, pattern, match.author = .BibOptions$match.author,
  ign.case = .BibOptions$ignore.case, regx = .BibOptions$use.regex)
}
\description{
Find name in a name field of a BibEntry object
}
\keyword{internal}

