% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[[<-.BibEntry}
\alias{[[<-.BibEntry}
\title{Assign a BibEntry entry to another BibEntry object}
\usage{
\method{[[}{BibEntry}(x, i) <- value
}
\arguments{
\item{x}{- a BibEntry object}

\item{i}{- a numeric index or a string entry key}

\item{value}{- a single entry BibEntry object or an object that can be
coereced to BibEntry using \code{\link{as.BibEntry}}}
}
\value{
an object of class BibEntry
}
\description{
Replace one entry in a BibEntry object with another
}
\details{
This function will replace the specified entry in \code{x} with the entry given
by \code{value}.  To replace multiple entries see \code{\link{[<-.BibEntry}}.
}
\seealso{
Other operators: \code{\link{$.BibEntry}};
  \code{\link{$<-.BibEntry}}; \code{\link{+.BibEntry}},
  \code{\link{+.BibEntry}}, \code{\link{merge.BibEntry}};
  \code{\link{SearchBib}}, \code{\link{[.BibEntry}},
  \code{\link{[.BibEntry}}; \code{\link{[<-.BibEntry}};
  \code{\link{[[.BibEntry}}; \code{\link{c.BibEntry}}
}
\keyword{methods}

