% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/levels.R
\name{levels.BibEntry}
\alias{fields}
\alias{levels.BibEntry}
\title{Extract all fields present in a BibEntry object}
\usage{
\method{levels}{BibEntry}(x)

fields(x)
}
\arguments{
\item{x}{a BibEntry object.}
}
\value{
a list with the same length as \code{x} of character vectors giving the
fields present in each entry of a BibEntry object.
}
\description{
These functions return a list of all fields present in a BibEntry object.
}
\note{
The only difference between \code{fields} and \code{levels} is that
\code{levels} returns a list with element names corresponding to entry keys.
}
\examples{
bib <- as.BibEntry(list(c(bibtype = "Article", key = "mclean2014a", title = "My New Article",
  author = "Mathew W. McLean",
  journaltitle = "The Journal", date = "2014-01"), c(bibtype = "Book", key = "mclean2014b",
  title = "My New Book", editor = "Mathew W. McLean", ISBN = "247123837", date = "2014-02")))
fields(bib)
levels(bib)
}
\keyword{methods}

